Test module dependencies in isolation
`grunt test-module-isolation` will run `karma:<module>` where `<module>`
is a module name (see folders in `src` and `dist`), running a test suite
of only that module's code `dist/<module>` and specs `tests/<module>`.
For now, the only specs are DRYed calls to `testModuleDependencies` in
`tests/helpers/test-module-dependencies.js` that loads the single module
and makes sure the injector doesn't throw an error.

This allows us to test if a module, when loaded in isolation, will run
without error.